Implement type prefix and suffix parsing in the `parser` package
- Introduce `parseTypePrefix` and `parseTypeSuffix` functions to handle type prefixes (e.g., `own`, `&`, `&mut`, `*`) and suffixes (e.g., `[]`, `[Expr]`), respectively.
- Support multiple prefixes and nested array types, enhancing the parser's ability to interpret complex type expressions.
- Improve diagnostic messages for missing brackets and array sizes, providing suggestions for corrections to enhance user experience.

These changes significantly enhance the type parsing capabilities of the Surge language, allowing for more expressive type definitions and better error handling during parsing.
